# docs/v3/api_spec/crs_converter_openapi.json
{
  "swagger": "2.0",
  "info": {
    "description": "Provides conversion/transformation services from a source to a target CRS. A coordinate operation is a conversion if the source and target CRS share the same datum; the operation is called transformation if datum transformations are involved, i.e. source and target CRS do not share the same datum.",
    "version": "1.0.0",
    "title": "CRS Conversion Service"
  },
  "host": "az-osdu1.evd.csp.slb.com",
...
  "basePath": "/api/crs/v1",
  "security": [
    {
      "Bearer": []
    },
    {
      "google_id_token": []
    }
  ],
...
  "schemes": [
    "https"
  ],
  "paths": {
...
    "/v3/convert": {
      "post": {
        "tags": [
          "CRS Point Conversion"
        ],
        "summary": "Convert a list of points",
        "description": "Convert a list of points",
        "operationId": "convertPoint",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": false,
            "schema": {
              "$ref": "#/definitions/ConvertPointsRequest"
            }
          },
          {
            "in": "header",
            "name": "data-partition-id",
            "required": false
          }
        ],

        "responses": {
          "200": {
            "description": "Conversion performed successfully",
            "schema": {
              "$ref": "#/definitions/ConvertPointsResponse"
            }
          },
          "400": {
            "description": "Bad input format",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "503": {
            "description": "CRS-converter overloaded; try again later. See HttpHeader 'Retry-After'.",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        },
        "security": [
          {
            "Bearer": []
          },
          {
            "google_id_token": []
          }
        ]
      }
    },
    "/v3/convertGeoJson": {
      "post": {
        "tags": [
          "CRS Point Conversion"
        ],
        "summary": "Convert a GeoJSON or AnyCrsGeoJson structure",
        "description": "Convert a GeoJSON or AnyCrsGeoJson structure to a specified target CRS. GeoJSON is declared - by definition - to be in the context of WGS 84.",
        "operationId": "convertGeoJson",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": false,
            "schema": {
              "$ref": "#/definitions/ConvertGeoJsonRequest"
            }
          },
          {
            "in": "header",
            "name": "data-partition-id",
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "Conversion performed successfully",
            "schema": {
              "$ref": "#/definitions/ConvertGeoJsonResponse"
            }
          },
          "400": {
            "description": "Bad input format",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "503": {
            "description": "CRS-converter overloaded; try again later. See HttpHeader 'Retry-After'.",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        },
        "security": [
          {
            "Bearer": []
          },
          {
            "google_id_token": []
          }
        ]
      }
    },
...
  },
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },
  "definitions": {
...
    "Point": {
      "type": "object",
      "required": [
        "x",
        "y",
        "z"
      ],
      "properties": {
        "x": {
          "type": "number",
          "format": "double",
          "example": -61.04340628871454,
          "description": "x coordinate or longitude"
        },
        "y": {
          "type": "number",
          "format": "double",
          "example": 10.673103179456877,
          "description": "y coordinate or latitude"
        },
        "z": {
          "type": "number",
          "format": "double",
          "example": 0,
          "description": "z coordinate"
        }
      },
      "description": "Point representation for CRS operations"
    },
...
    "ErrorResponse": {
      "type": "object",
      "properties": {
        "error": {
          "type": "string",
          "description": "Error message"
        }
      },
      "description": "Response resulting of a failed CRS convertion operation"
    },
...
    "ConvertPointsResponse": {
      "type": "object",
      "properties": {
        "successCount": {
          "type": "integer",
          "format": "int32",
          "description": "Number of points successfully converted. If the number is less than the request array length conversion/transformation failures occurred."
        },
        "points": {
          "type": "array",
          "description": "Converted points; length and order of the array is the same as in the request. Points, which failed to convert, are returned as NaN.",
          "items": {
            "$ref": "#/definitions/Point"
          }
        },
        "operationsApplied": {
          "type": "array",
          "description": "The list of operations, which have been applied to the points",
          "items": {
            "type": "string"
          }
        }
      },
      "description": "Response of a CRS conversion/transformation operation"
    },
    "ConvertPointsRequest": {
      "type": "object",
      "required": [
        "fromCRS",
        "points",
        "toCRS"
      ],
      "properties": {
        "fromCRS": {
          "type": "string",
          "example": "osdu:reference-data--CoordinateReferenceSystem:BoundCRS::osdu::fromcrsConvert23031024",
          "description": "Source CRS as persistable reference string or a record ID"
        },
        "toCRS": {
          "type": "string",
          "example": "osdu:reference-data--CoordinateReferenceSystem:BoundCRS::osdu::toCRSconvert23031024",
          "description": "Target CRS as persistable reference string or a record ID"
        },
        "points": {
          "type": "array",
          "description": "List of points to be converted",
          "items": {
            "$ref": "#/definitions/Point"
          }
        }
      },
      "description": "Request to convert a set of points from a source CRS to a target CRS.",
      "example": {
        "fromCRS": "osdu:reference-data--CoordinateReferenceSystem:BoundCRS::osdu::fromcrsConvert23031024",
        "toCRS":"osdu:reference-data--CoordinateReferenceSystem:BoundCRS::osdu::toCRSconvert23031024",
        "points": [
          {
            "x": -61.04340628871454,
            "y": 10.673103179456877,
            "z": 0
          }
        ]
      }
    },
    "ConvertGeoJsonRequest": {
      "title": "Convert GeoJson/AnyCrsGeoJson Request",
      "description": "Request to convert a GeoJSON FeatureCollection or AnyCrsFeatureCollection from WGS 84 or AnyCrsFeatureCollection.persistableReferenceCrs to a target CRS.",
      "type": "object",
      "required": [
        "featureCollection",
        "toCRS"
      ],
      "properties": {
        "featureCollection": {
          "$ref": "#/definitions/AnyCrsGeoJsonFeatureCollection"
        },
        "toCRS": {
          "title": "GeoJson Request",
          "type": "string",
          "description": "The GeoJSON FeatureCollection or AnyCrsFeatureCollection structure to be converted/transformed. GeoJSON is always based on WGS 84; AnyCrsFeatureCollection carries the CRS context in the persistableReferenceCrs property. This value could be persistable reference string or a record ID"
        },
        "toUnitZ": {
          "type": "string",
          "description": "The vertical axis unit for this AnyCrsGeoJsonFeatureCollection as persistable reference string or a record ID)",
          "title": "Vertical Unit"
        }
      },
      "example": {
        "toCRS": "osdu:reference-data--CoordinateReferenceSystem:BoundCRS::osdu::toCRSGeojson23031024",
        "toUnitZ": "osdu:reference-data--UnitOfMeasure:ms",
        "featureCollection": {
          "features": [
            {
              "geometry": {
                "coordinates": [
                  313405.9477893702,
                  6544797.620047403,
                  6.561679790026246
                ],
                "bbox": null,
                "type": "AnyCrsPoint"
              },
              "bbox": null,
              "properties": {},
              "type": "AnyCrsFeature"
            }
          ],
          "bbox": null,
          "properties": {},
          "CoordinateReferenceSystemID": "osdu:reference-data--CoordinateReferenceSystem:BoundCRS::osdu::persistGeojson23031024",
          "VerticalUnitZ": "osdu:reference-data--UnitOfMeasure:ms",
          "type": "AnyCrsFeatureCollection"
        }
      }
    },
    "ConvertGeoJsonResponse": {
      "title": "Convert GeoJSON/AnyCrsGeoJson response",
      "description": "Response of a CRS conversion/transformation operation involving GeoJSON FeatureCollection or AnyCrsFeatureCollection.",
      "type": "object",
      "properties": {
        "totalCount": {
          "type": "integer",
          "description": "The total number of coordinates in the GeoJSON FeatureCollection or AnyCrsFeatureCollection."
        },
        "successCount": {
          "type": "integer",
          "description": "The number of coordinates in the GeoJSON FeatureCollection or AnyCrsFeatureCollection successfully converted/transformed. If this number is less than totalCount then conversion/transformation errors have occurred."
        },
        "featureCollection": {
          "$ref": "#/definitions/AnyCrsGeoJsonFeatureCollection"
        },
        "operationsApplied": {
          "type": "array",
          "description": "The list of operations, which have been applied to the points",
          "items": {
            "type": "string"
          }
        }
      },
      "example": {
        "successCount": 1,
        "totalCount": 1,
        "featureCollection": {
          "features": [
            {
              "geometry": {
                "coordinates": [
                  5.750000000484756,
                  59.00000000018591,
                  2
                ],
                "bbox": null,
                "type": "Point"
              },
              "bbox": null,
              "properties": {},
              "type": "Feature"
            }
          ],
          "bbox": null,
          "properties": {},
          "persistableReferenceCrs": null,
          "type": "FeatureCollection"
        },
        "operationsApplied": [
          "conversion from ED_1950_UTM_Zone_32N to GCS_European_1950; 1 points converted",
          "transformation GCS_European_1950 to GCS_WGS_1984 using ED_1950_To_WGS_1984_23; 1 points successfully transformed",
          "Z-axis unit conversion from ft to m"
        ]
      }
    },
...
    "AnyCrsGeoJsonFeatureGeometry": {
      "type": "object",
      "description": "AnyCrsGeoJSon feature geometry",
      "discriminator": "type",
      "required": [
        "type"
      ],
      "properties": {
        "type": {
          "type": "string",
          "enum": [
            "Point",
            "LineString",
            "Polygon",
            "MultiPoint",
            "MultiLineString",
            "MultiPolygon",
            "GeometryCollection",
            "AnyCrsPoint",
            "AnyCrsLineString",
            "AnyCrsPolygon",
            "AnyCrsMultiPoint",
            "AnyCrsMultiLineString",
            "AnyCrsMultiPolygon",
            "AnyCrsGeometryCollection"
          ],
          "description": "the geometry type"
        },
        "coordinates": {
          "type": "object"
        }
      }
    },
    "AnyProperty": {
      "title": "Any property extension",
      "type": "object",
      "additionalProperties": {
        "type": "object"
      }
    },
    "AnyCrsGeoJsonFeature": {
      "title": "AnyCrsGeoJSON Feature",
      "type": "object",
      "required": [
        "type",
        "properties",
        "geometry"
      ],
      "properties": {
        "type": {
          "type": "string",
          "enum": [
            "Feature",
            "AnyCrsFeature"
          ]
        },
        "properties": {
          "$ref": "#/definitions/AnyProperty"
        },
        "geometry": {
          "$ref": "#/definitions/AnyCrsGeoJsonFeatureGeometry"
        },
        "bbox": {
          "type": "array",
          "minItems": 4,
          "items": {
            "type": "number"
          }
        }
      }
    },
    "AnyCrsGeoJsonFeatureCollection": {
      "title": "AnyCrsGeoJSON FeatureCollection",
      "type": "object",
      "required": [
        "type",
        "features"
      ],
      "properties": {
        "type": {
          "type": "string",
          "enum": [
            "FeatureCollection",
            "AnyCrsFeatureCollection"
          ]
        },
        "features": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/AnyCrsGeoJsonFeature"
          }
        },
        "bbox": {
          "type": "array",
          "minItems": 4,
          "items": {
            "type": "number"
          }
        },
        "persistableReferenceCrs": {
          "type": "string",
          "description": "The spatial context of this AnyCrsGeoJsonFeatureCollection as persistable reference string.",
          "title": "Spatial Reference"
        },
        "CoordinateReferenceSystemID": {
          "type": "string",
          "description": "The spatial context of this AnyCrsGeoJsonFeatureCollection as a CoordinateReferenceSystemID.",
          "title": "Spatial Reference"
        },
        "VerticalUnitZ": {
          "type": "string",
          "description": "The vertical axis unit for this AnyCrsGeoJsonFeatureCollection as a record ID.",
          "title": "Vertical Unit"
        },
        "persistableReferenceUnitZ": {
          "type": "string",
          "description": "The vertical axis unit for this AnyCrsGeoJsonFeatureCollection as persistable reference string.",
          "title": "Vertical Unit"
        }
      }
    },
...
  }
}